requirements.txt
humanize==2.3.0
peewee==3.13.2
python-dotenv==0.13.0
requests==2.23.0